drm: rcar-du: Fix reflected method of DPTSR register
The processing which returns a setup of DU1 to an initial value
was performed at the time of the processing which makes an overlay
layer undisplayed. The DPTSR register was then updated.
When an overlay layer is made undisplayed,
a setup of DPTSR register is not changed by applying this patch.
The DPTSR register is not updated unless the output channel of
overlay layer is changed.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>