usb: renesas_usbhs: fix spinlock lockup suspected in a gadget driver
The usbhsf_pkt_handler(pipe, USBHSF_PKT_DMA_DONE) in usbhsf_dma_complete()
will call the complete function of a usb gadget driver finally.
According to the gadget.h, the complete function assumes that it is
called with interrupts disabled. However, the shdmac driver will call
the usbhsf_dma_complete() as the callback in tasklet.

So, this patch adds a local_irq_save/local_irq_restore in the
usbhsf_dma_complete(). And then, the spinlock lockup suspected issue
disappeared.

Signed-off-by: Yoshihiro Shimoda <yoshihiro.shimoda.uh@renesas.com>