usb: renesas_usbhs: Fix fifo unclear in usbhsf_prepare_pop
This patch fixes an issue of usb gadget with f_audio_source.
When usbhsf_prepare_pop() is called after audio_setup() in f_audio_source,
usb host does not write in fifo.
A setup of usb host is not completed and the error message of
"cannot set freq 44100 to ep 0x81" is output.
This patch adds fifo clear in usbhsf_prepare_pop().

Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>