OF: Introduce helper function for getting PCI domain_nr
Add of_pci_get_domain_nr() to retrieve the PCI domain number
of a given device from DT. If the information is not present,
the function can be requested to allocate a new domain number.

Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Grant Likely <grant.likely@linaro.org>
Cc: Rob Herring <robh+dt@kernel.org>
Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>

This is a backport of a patch sent on the PCI mailing list but
not yet accepted upstream. Additionally, fixed a bug in the
of_pci_get_domain_nr() code:
-	if (of_domain_nr == -1) {
+	if (domain == -1) {

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>