drm: rcar-du: Fix redundant reset of Display Unit
When reflecting a value in DEFR8 register, it is necessary to
perform DRES (Display Reset). Even if the value of DEFR8 register
is not changed, the current code has executed DRES. Therefore,
when DEFR8 register was always set up, the screen had disappeared
for a moment. This patch does not execute DRES, when not changing
DEFR8 register. Moreover, bit write restriction in DEFR8 register
is added for every devices.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>