ASoC: rcar: Fix audio underrun error/status check failure
This patch fixes an issue with audio DMA underrun occuring before
the start of the first SSI transfer. Prior to this fix, the audio
DMA is started before starting the SSI. If the SSI is not started
quick enough, you will get SSI buffer full errors. This only shows
up when used on a heavily loaded system.

The fix is simply to follow the guideline in the User Manual, i.e.
set the SSIq/r_CONTROL.start bit after setting SSICR.en bit.

Signed-off-by: Biju Das <biju.das@bp.renesas.com>
Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>