ARM: shmobile: koelsch-reference: Fix buswait_bwait for HSUSB hp_clk
The buswait_bwait is the number of access waits to HSUSB.
When the buswait_bwait is 0, it is 2 cycle.
When the buswait_bwait is 9, it is 11 cycle.
The cycle of continuous accesses to the registers of HSUSB
must be 80 ns or more.
The hp_clk is at most 130 MHz. So buswait_bwait is set 9.

Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>