drm: rcar-du: Fix up/down scale on interlace
When the plane is scaled down and scaled up at the time of interlace
mode, a problem that a image is flickered is corrected.
About detail, for an UDS(Up Down Scale) to have to refer to
a setting parameter of a progressive image. it was collected so that
the IP conversion parameter which was set by set_plane_param() was
performed in the vspd_drv_main.c:check_rpf_param().

Signed-off-by: Takashi Saito <takashi.saitou.ry@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>