drm: rcar-du: Fix RPF display position on interlaced mode
The RPF (Read Pixel Formatter) vertical display position parameter is
needed to set to 1/2 on interlaced mode. But, in the case of a scaling
path, the vertical display position have been not set to 1/2. By this
patch, it was corrected the vertical display position to 1/2 in the case
of the scaling path.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>