dma: rcar-dma: hack: Map the DMA slave addresses
Without an IOMMU the phys_addr_t and dma_addr_t are
the same and no special care is needed.
With an IOMMU we have to map the DMA slave phys_addr_t to a dma_addr_t.
The purpose of this hack is to do 1:1 mapping by hand while dmaengine
_doesn't_ care about it.

This hack is only needed for the IOMMU test cases until the proper solution
reaches upstream.
Tested with aplay tool where this connection is involved:
RCAR_SOUND_IPs->AUDMACs->IPMMU_M1->MEMORY

Here is one more hack how to solve the similar issue, but it is not suitaible
for us:
https://patchwork.kernel.org/patch/9341527/

The proper solution is based on v4.8:
https://www.mail-archive.com/linux-renesas-soc@vger.kernel.org/msg06395.html