ARM: mach-shmobile: Don't init CNTVOFF if PSCI is available
If PSCI is available then most likely we are running on PSCI-enabled
U-Boot which, we assume, has already taken care of resetting CNTVOFF
before switching to non-secure mode and we don't need to.

Also, don't init CNTVOFF if we are running on top of Xen hypervisor,
as CNTVOFF is controlled by hypervisor itself and shouldn't be touched
by Dom0 in such case.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
CC: Julien Grall <julien.grall@arm.com>

---
   You can find previous discussion here:
   https://lkml.org/lkml/2019/4/17/810

   Changes in v2:
      - Clarify patch subject/description
      - Don't use CONFIG_ARM_PSCI option, check whether the PSCI is available,
        by using psci_smp_available()
      - Check whether we are running on top of Xen, by using xen_domain()
---