dt-bindings: xen: Add xen,dev-domid property description for xen-virtio layer
Introduce Xen specific binding for the virtio device to be used by
Xen virtio support driver in a subsequent commit.

This binding specifies the ID of Xen domain where the corresponding
device (backend) resides. This is needed for the option to restrict
memory access using Xen grant mappings to work.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>